prompting_lambda/requirements.txt
google-generativeai
testresources
